providers/proxy: fix URL path getting lost when partial URL is given to rd=
